GRHB-631: * split interviews navigation
