+++ mobile/src/navigation/app/common/constants/constants.ts
@@ -11,7 +11,7 @@ import { Settings } from '~/components/setting/setting';
 import { Chat } from '~/navigation/chat/chat.navigation';
 import { Courses } from '~/navigation/courses/courses.navigation';
 import { CoursesManagement } from '~/navigation/courses-management/courses-management.navigation';
-import { Interviews } from '~/navigation/interview/interview-stack.navigation';
+import { Interviews } from '~/navigation/interviews/interviews.navigation';
 import { MyCourses } from '~/navigation/my-courses/my-courses-stack.navigation';
 import { UAM } from '~/navigation/uam/uam.navigation';
