+++ mobile/src/navigation/interview/common/constants.ts
@@ -1,28 +1,17 @@
 import { MaterialTopTabNavigationOptions } from '@react-navigation/material-top-tabs';
-import { NativeStackNavigationOptions } from '@react-navigation/native-stack';
 
 import {
   AppColor,
   AppFontFamily,
   InterviewScreenName,
-  InterviewsScreenName,
-  PermissionKey,
 } from '~/common/enums/enums';
 import { NavigationItem } from '~/common/types/types';
 import {
   Applications,
   Notes,
 } from '~/components/interview/components/components';
-import { Interviews } from '~/components/interviews/interviews';
 
-import { Interview } from '../interview.navigation';
-
-const SCREEN_OPTIONS: NativeStackNavigationOptions = {
-  headerTitleAlign: 'center',
-  headerShown: false,
-};
-
-const TAB_OPTIONS: MaterialTopTabNavigationOptions = {
+const SCREEN_OPTIONS: MaterialTopTabNavigationOptions = {
   tabBarLabelStyle: {
     fontFamily: AppFontFamily.INTER_600,
     fontSize: 14,
@@ -41,7 +30,7 @@ const TAB_OPTIONS: MaterialTopTabNavigationOptions = {
   tabBarInactiveTintColor: AppColor.TEXT.GRAY_200,
 };
 
-const INTERVIEW_TAB_ITEMS: NavigationItem[] = [
+const NAVIGATION_ITEMS: NavigationItem[] = [
   {
     name: InterviewScreenName.APPLICATIONS,
     component: Applications,
@@ -56,22 +45,4 @@ const INTERVIEW_TAB_ITEMS: NavigationItem[] = [
   },
 ];
 
-const NAVIGATION_ITEMS: NavigationItem[] = [
-  {
-    name: InterviewsScreenName.INTERVIEWS,
-    component: Interviews,
-    isAuthRequired: true,
-    permissions: [
-      PermissionKey.MANAGE_INTERVIEW,
-      PermissionKey.MANAGE_INTERVIEWS,
-    ],
-  },
-  {
-    name: InterviewsScreenName.INTERVIEW,
-    component: Interview,
-    isAuthRequired: true,
-    permissions: [],
-  },
-];
-
-export { INTERVIEW_TAB_ITEMS, NAVIGATION_ITEMS, SCREEN_OPTIONS, TAB_OPTIONS };
+export { NAVIGATION_ITEMS, SCREEN_OPTIONS };
