+++ mobile/src/navigation/interview/interview.navigation.tsx
@@ -7,13 +7,13 @@ import { InterviewNavigationParamList } from '~/common/types/types';
 import { BackButton, View } from '~/components/common/common';
 import { useAppNavigate } from '~/hooks/hooks';
 
-import { INTERVIEW_TAB_ITEMS, TAB_OPTIONS } from './common/constants';
+import { NAVIGATION_ITEMS, SCREEN_OPTIONS } from './common/constants';
 import { styles } from './styles';
 
 const Tab = createMaterialTopTabNavigator<InterviewNavigationParamList>();
 
 const Interview: FC = () => {
-  const isTabsShown = INTERVIEW_TAB_ITEMS.length > MIN_SCREENS_COUNT_FOR_TABS;
+  const isTabsShown = NAVIGATION_ITEMS.length > MIN_SCREENS_COUNT_FOR_TABS;
   const navigation = useAppNavigate();
 
   useEffect(() => {
@@ -28,8 +28,8 @@ const Interview: FC = () => {
 
   return (
     <View style={styles.container}>
-      <Tab.Navigator screenOptions={TAB_OPTIONS}>
-        {INTERVIEW_TAB_ITEMS.map((screen) => (
+      <Tab.Navigator screenOptions={SCREEN_OPTIONS}>
+        {NAVIGATION_ITEMS.map((screen) => (
           <Tab.Screen
             key={screen.name}
             name={screen.name as InterviewScreenName}
